+++ builder_pkgs/mockito/example/iss/iss_test.mocks.dart
@@ -17,6 +17,7 @@ import 'iss.dart' as _i4;
 // ignore_for_file: comment_references
 // ignore_for_file: deprecated_member_use
 // ignore_for_file: deprecated_member_use_from_same_package
+// ignore_for_file: experimental_member_use
 // ignore_for_file: implementation_imports
 // ignore_for_file: invalid_use_of_visible_for_testing_member
 // ignore_for_file: must_be_immutable