+++ builder_pkgs/mockito/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 * Suppress `experimental_member_use` warnings in generated mock files.
   [#4602](https://github.com/dart-lang/build/issues/4602)
+* Document missing stub behavior for `void` and `Future<void>`.
 
 ## 5.8.0
+
 * Add support for JS interop extension type mock generation to mockito.
 * It is now an error to generate mocks for non-JS interop extension types.
   Instead, mock the representation type and cast to the extension type where