+++ builder_pkgs/mockito/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 5.8.1-wip
 
+* Suppress `experimental_member_use` warnings in generated mock files.
+  [#4602](https://github.com/dart-lang/build/issues/4602)
 * Document missing stub behavior for `void` and `Future<void>`.
 
 ## 5.8.0