+++ builder_pkgs/mockito/CHANGELOG.md
@@ -3,7 +3,12 @@
 * Remove redundant per-class `// ignore: must_be_immutable` comment to prevent
   `duplicate_ignore` warnings.
   [#4588](https://github.com/dart-lang/build/issues/4588)
-
+* Clarify in `README.md` that generated `.mocks.dart` filenames are based on the
+  source file name.
+  [#4599](https://github.com/dart-lang/build/issues/4599)
+* Suppress `experimental_member_use` warnings in generated mock files.
+  [#4602](https://github.com/dart-lang/build/issues/4602)
+* Document missing stub behavior for `void` and `Future<void>`.
 ## 5.8.0
 
 * Add support for JS interop extension type mock generation to mockito.