+++ builder_pkgs/mockito/CHANGELOG.md
@@ -13,6 +13,7 @@
 * Suppress `experimental_member_use` warnings in generated mock files.
   [#4602](https://github.com/dart-lang/build/issues/4602)
 * Document missing stub behavior for `void` and `Future<void>`.
+
 ## 5.8.0
 
 * Add support for JS interop extension type mock generation to mockito.